// src/app/home/types/order.type.ts
export interface OrderItem {
  productId: number;
  qty: number;
  price: number;
  amount: number;
}

export interface Order {
  userName: string;
  address: string;
  city: string;
  state: string;
  pin: string;
  total: number;
  userEmail: string;
  orderDetails: OrderItem[];
}

export interface PastOrder {
  userName: string;
  address: string;
  city: string;
  state: string;
  pin: string;
  total: number;
  orderDate: string;
  orderId: number;
}

export interface PastOrderProduct {
  productId: number;
  productImg: string;
  qty: number;
  price: number;
  amount: number;
  productName: string;
}